+++ container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend/docker_kurtosis_backend_module_functions.go
@@ -1,4 +1,4 @@
-package docker
+package docker_kurtosis_backend
 
 import (
 	"context"
@@ -33,6 +33,8 @@ const (
 	maxWaitForModuleContainerAvailabilityRetries         = 10
 	timeBetweenWaitForModuleContainerAvailabilityRetries = 1 * time.Second
 )
+
+// TODO: MIGRATE THIS FOLDER TO USE STRUCTURE OF USER_SERVICE_FUNCTIONS MODULE
 
 func (backend *DockerKurtosisBackend) CreateModule(
 	ctx context.Context,