+++ container-engine-lib/docs/changelog.md
@@ -1,5 +1,23 @@
 # TBD
 
+#### Breaking Changes
+* Move all `docker.DockerKurtosisBackend` functionality to `docker_kurtosis_backend` module
+  * Users will need to replace their `github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker` import for
+`github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend`
+* Move all functions in `DockerKurtosisBackend` related to user services to `user_service_functions` module
+  * Users will need to replace their `github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker` import for
+`github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend`
+* Move all `kubernetes.KubernetesKurtosisBackend` functionality to `kubernetes_kurtosis_backend` module
+  * Users will need to replace their `github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/kubernetes` import for
+`github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/kubernetes/kubernetes_kurtosis_backend`
+* Move all functions in `KubernetesKurtosisBackend` related to user services to `user_service_functions` module
+  * Users will need to replace their `github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/kubernetes` import for
+`github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/kubernetes/kubernetes_kurtosis_backend`
+* Move `ServiceConfig` and `FilesArtifactsExpansion` out of `backend_interface` module into respective files in `objects` module
+  * Users will need to replace their `github.com/kurtosis-tech/container-engine-lib/lib/backend_interface` import for `ServiceConfig` 
+and `FilesArtifactsExpansion` to be `github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/service` 
+and `github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/files_artifacts_expansion` respectively
+
 ### Changes
 * Add `RegisterUserServices` and `StartUserServices` to `KurtosisBackend` interface
 * Implement `RegisterUserServices` in `DockerKurtosisBackend`